test(125): extract cross-cutting variants into modules/ui/variant/
Creates `modules/ui/variant/` as the home for presentation-modifier classes — opt-in styling
utilities that any component can mix into its className. Moves what already existed:

- `misc/Color.*` → `variant/Color.*`
- `misc/Status.*` → `variant/Status.*`
- `misc/Surface.*` → `variant/Surface.*`
- `block/Flex.*` → `variant/Flex.*`

Adds two new utilities that were previously buried as component-internal variants:

- `variant/Align.*` — `.left`, `.center`, `.right` text-alignment, `getAlignClass()`
- `variant/Spacing.*` — `.spacious` margin tweak, `getSpacingClass()`

Migrates `Paragraph` and `Block` to compose these utilities: their `text-align` default and
`.spacious` variant rules delete, the alignment/spacing props extend the new variant interfaces.
The `:where(&) { text-align: left }` low-specificity workaround in `Paragraph.module.css` is no
longer needed (no default to suppress) and goes away.

Audits `Flex` to be a "dumb box" — drops the implicit `BLOCK_CLASS` wrap on the `<Flex>` component
so it carries no external margin, and formalises `getFlexClass(variants)` as the public API
(matching `getColorClass` / `getStatusClass`). Other components that need flex layout continue to
use `FLEX_CSS` directly or migrate to `getFlexClass`.

Sweeps 20 importers across `modules/ui/{block,form,inline,misc,notice}/` and `modules/ui/docs/`.
The `misc/` and `block/` barrels lose the moved exports; `ui/index.ts` gains a `variant/` export.

Verified across Chromium and Firefox: visual parity with the previous state, no regressions in the
stress test.
