+++ modules/ui/block/Card.tsx
@@ -1,9 +1,9 @@
 import type { ReactElement } from "react";
 import { Clickable, type ClickableProps } from "../form/Clickable.js";
-import { type ColorVariants, getColorClass } from "../misc/Color.js";
-import { getStatusClass, type Status } from "../misc/Status.js";
-import { SURFACE_CLASS } from "../misc/Surface.js";
 import { getClass, getModuleClass } from "../util/css.js";
+import { type ColorVariants, getColorClass } from "../variant/Color.js";
+import { getStatusClass, type Status } from "../variant/Status.js";
+import { SURFACE_CLASS } from "../variant/Surface.js";
 import CARD_CSS from "./Card.module.css";
 
 export interface CardProps extends ClickableProps, ColorVariants {
