+++ modules/ui/block/Paragraph.module.css
@@ -1,30 +1,10 @@
 .paragraph,
 .prose p {
-	/* Low-specificity defaults */
-	:where(&) {
-		/* Typography */
-		text-align: left;
-	}
-
 	/* Box */
 	display: block;
 	inline-size: 100%;
 	margin-inline: 0;
 	margin-block: var(--paragraph-spacing, var(--space-normal));
-
-	/* Variants */
-	&.left {
-		text-align: left;
-	}
-	&.center {
-		text-align: center;
-	}
-	&.right {
-		text-align: right;
-	}
-	&.spacious {
-		margin-block: var(--paragraph-spacious-spacing, var(--space-xxlarge));
-	}
 
 	/* Psuedo-classes */
 	&:first-child {
