+++ modules/ui/block/Paragraph.tsx
@@ -1,21 +1,24 @@
 import type { ReactElement } from "react";
-import { getModuleClass } from "../util/css.js";
+import { getClass, getModuleClass } from "../util/css.js";
 import type { OptionalChildProps } from "../util/props.js";
+import { type AlignVariants, getAlignClass } from "../variant/Align.js";
+import { getSpacingClass, type SpacingVariants } from "../variant/Spacing.js";
 import PARAGRAPH_CSS from "./Paragraph.module.css";
 
-export interface ParagraphProps extends OptionalChildProps {
-	/** Align the paragraph to the right. */
-	right?: boolean | undefined;
-	/** Align the paragraph to the center. */
-	center?: boolean | undefined;
-	/** Align the paragraph to the left. */
-	left?: boolean | undefined;
-	/** Give the paragraph more space. */
-	spacious?: boolean | undefined;
-}
+export interface ParagraphProps extends AlignVariants, SpacingVariants, OptionalChildProps {}
 
 export function Paragraph({ children, ...variants }: ParagraphProps): ReactElement {
-	return <p className={getModuleClass(PARAGRAPH_CSS, "paragraph", variants)}>{children}</p>;
+	return (
+		<p
+			className={getClass(
+				getModuleClass(PARAGRAPH_CSS, "paragraph"), //
+				getAlignClass(variants),
+				getSpacingClass(variants),
+			)}
+		>
+			{children}
+		</p>
+	);
 }
 
 export { PARAGRAPH_CSS };
