+++ modules/ui/form/ArrayInput.tsx
@@ -4,7 +4,7 @@ import type { Schema } from "../../schema/Schema.js";
 import { type ImmutableArray, omitArrayIndex, withArrayIndex } from "../../util/array.js";
 import { splitMessage } from "../../util/error.js";
 import { formatUnit } from "../../util/format.js";
-import { Flex } from "../block/Flex.js";
+import { Flex } from "../variant/Flex.js";
 import { Button } from "./Button.js";
 import { ButtonInput } from "./ButtonInput.js";
 import type { ValueInputProps } from "./Input.js";
