+++ modules/ui/misc/Tag.tsx
@@ -1,9 +1,9 @@
 import type { ReactElement } from "react";
 import { Clickable, type ClickableProps } from "../form/Clickable.js";
 import { getClass, getModuleClass } from "../util/css.js";
-import { type ColorVariants, getColorClass } from "./Color.js";
-import { getStatusClass, type StatusVariants } from "./Status.js";
-import { SURFACE_CLASS } from "./Surface.js";
+import { type ColorVariants, getColorClass } from "../variant/Color.js";
+import { getStatusClass, type StatusVariants } from "../variant/Status.js";
+import { SURFACE_CLASS } from "../variant/Surface.js";
 import TAG_CSS from "./Tag.module.css";
 
 /** Variants for tags — accepts both status and raw colour variants. */
