+++ modules/ui/misc/index.tsx
@@ -1,11 +1,8 @@
 export * from "./Catcher.js";
-export * from "./Color.js";
 export * from "./Loading.js";
 export * from "./Mapper.js";
 export * from "./Markup.js";
 export * from "./MetaContext.js";
-export * from "./Status.js";
 export * from "./StatusIcon.js";
-export * from "./Surface.js";
 export * from "./Tag.js";
 export * from "./Typography.js";
