+++ modules/ui/variant/Align.module.css
@@ -1,0 +1,13 @@
+/* Text-alignment variants — apply via `getAlignClass()` to any component that wants opt-in alignment. */
+
+.left {
+	text-align: left;
+}
+
+.center {
+	text-align: center;
+}
+
+.right {
+	text-align: right;
+}
